--- src/main/java/gwangjang/server/domain/application/dto/res/MemberDto.java
@@ -1,18 +1,0 @@
-package gwangjang.server.domain.application.dto.res;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Setter
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor
-public class MemberDto {
-
-    private String memberId;
-    private String nickname;;
-    private String profileImage;;
-
-}
