--- src/main/java/gwangjang/server/domain/application/dto/res/MemberRes.java
@@ -1,8 +1,0 @@
-package gwangjang.server.domain.application.dto.res;
-
-public class MemberRes {
-
-    private String nickname;
-    private String profileImg;
-
-}
