+++ src/main/java/gwangjang/server/domain/community/application/dto/req/CommunityReq.java
@@ -1,0 +1,18 @@
+package gwangjang.server.domain.community.application.dto.req;
+
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+public class CommunityReq {
+
+    private String title;
+    private String talk;
+
+}
