+++ src/main/java/gwangjang/server/domain/community/application/dto/res/MemberRes.java
@@ -1,0 +1,8 @@
+package gwangjang.server.domain.community.application.dto.res;
+
+public class MemberRes {
+
+    private String nickname;
+    private String profileImg;
+
+}
