+++ src/main/java/gwangjang/server/domain/community/domain/repository/CommunityRepository.java
@@ -1,0 +1,9 @@
+package gwangjang.server.domain.community.domain.repository;
+
+import gwangjang.server.domain.community.domain.entity.Community;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CommunityRepository extends JpaRepository<Community, Long>, CommunityCustomRepository {
+
+
+}
