+++ src/main/java/gwangjang/server/domain/community/exception/CommunityException.java
@@ -1,0 +1,12 @@
+package gwangjang.server.domain.community.exception;
+
+import gwangjang.server.global.exception.ApplicationException;
+import gwangjang.server.global.response.ErrorCode;
+import org.springframework.http.HttpStatus;
+
+public abstract class CommunityException  extends ApplicationException {
+
+    protected CommunityException(ErrorCode errorCode, HttpStatus httpStatus) {
+        super(errorCode, httpStatus);
+    }
+}
