--- src/main/java/gwangjang/server/domain/domain/repository/CommunityRepository.java
@@ -1,9 +1,0 @@
-package gwangjang.server.domain.domain.repository;
-
-import gwangjang.server.domain.domain.entity.Community;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CommunityRepository extends JpaRepository<Community, Long>, CommunityCustomRepository {
-
-
-}
