--- src/main/java/gwangjang/server/domain/exception/CommunityException.java
@@ -1,12 +1,0 @@
-package gwangjang.server.domain.exception;
-
-import gwangjang.server.global.exception.ApplicationException;
-import gwangjang.server.global.response.ErrorCode;
-import org.springframework.http.HttpStatus;
-
-public abstract class CommunityException  extends ApplicationException {
-
-    protected CommunityException(ErrorCode errorCode, HttpStatus httpStatus) {
-        super(errorCode, httpStatus);
-    }
-}
